fix register

+++ src/domain/user.rs
@@ -7,7 +7,7 @@ use argon2::{
 };
 
 pub struct User {
-    pub id: Option<i64>,
+    pub id: Option<i32>,
     pub username: String,
     pub first_name: Option<String>,
     pub last_name: Option<String>,
